moving landing page to discover/
