+++ pages/discover/index.tsx
