+++ screens/discover/_components/AcceptInviteModal.tsx
@@ -11,7 +11,7 @@ import getErrorMessage from 'src/utils/errorUtils'
 import { getExplorerUrlForTxHash } from 'src/utils/formattingUtils'
 import { addAuthorizedUser } from 'src/utils/gaslessUtils'
 import { delay } from 'src/utils/generics'
-import { InviteInfo, useJoinInvite } from 'screens/_utils/invite'
+import { InviteInfo, useJoinInvite } from 'screens/discover/_utils/invite'
 import logger from 'src/utils/logger'
 import { ForwardArrow } from 'src/v2/assets/custom chakra icons/Arrows/ForwardArrow'
 import RoleDataDisplay from 'src/v2/components/AcceptInviteModal/RoleDataDisplay'
