+++ screens/discover/_hooks/useMultiChainPaginatedQuery.ts
