+++ screens/discover/index.tsx
@@ -4,14 +4,14 @@ import { Container, Divider, Flex, HStack, Text, useToast } from '@chakra-ui/rea
 import { useRouter } from 'next/router'
 
 import { WebwalletContext } from 'pages/_app' //TODO - move to /libraries/zero-wallet/context
-import DaosGrid from 'screens/_components/DaosGrid'
+import DaosGrid from 'screens/discover/_components/DaosGrid'
 import { GetDaOsForExploreQuery, useGetDaOsForExploreQuery, Workspace_Filter as WorkspaceFilter, Workspace_OrderBy as WorkspaceOrderBy } from 'src/generated/graphql'
-import { useMultichainDaosPaginatedQuery } from 'screens/_hooks/useMultiChainPaginatedQuery'
+import { useMultichainDaosPaginatedQuery } from 'screens/discover/_hooks/useMultiChainPaginatedQuery'
 import NavbarLayout from 'libraries/ui/navbarLayout'
-import { extractInviteInfo, InviteInfo } from 'screens/_utils/invite'
+import { extractInviteInfo, InviteInfo } from 'screens/discover/_utils/invite'
 import logger from 'libraries/logger'
-import { mergeSortedArrays } from 'screens/_utils/mergeSortedArrays'
-import AcceptInviteModal from 'screens/_components/AcceptInviteModal'
+import { mergeSortedArrays } from 'screens/discover/_utils/mergeSortedArrays'
+import AcceptInviteModal from 'screens/discover/_components/AcceptInviteModal'
 
 const PAGE_SIZE = 3
